api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.29.7
python-dotenv==1.0.0
azure-storage-blob==12.27.1
assemblyai==0.25.0
anthropic==0.39.0
python-multipart==0.0.6
aiofiles==23.2.1
